Some fixes for dart2wasm (flutter/engine#38167)
* A few fixes to get things compiling via dart2wasm.

* Use `futureToPromise`.

* Some more fixes for dart2wasm.

* Allow promises to have nullable objects.

+++ engine/src/flutter/lib/web_ui/lib/src/engine/canvaskit/surface.dart
@@ -325,7 +325,7 @@ class Surface {
           // Default to no anti-aliasing. Paint commands can be explicitly
           // anti-aliased by setting their `Paint` object's `antialias` property.
           antialias: _kUsingMSAA ? 1 : 0,
-          majorVersion: webGLVersion,
+          majorVersion: webGLVersion.toDouble(),
         ),
       ).toInt();
